refactor: rename TwitchAlert folder for pep8
