+++ tests/cogs/TwitchAlert/test_twitch_handler.py
@@ -1,5 +1,5 @@
-from koala.cogs.TwitchAlert.twitch_handler import TwitchAPIHandler
-from koala.cogs.TwitchAlert.utils import TWITCH_KEY, TWITCH_SECRET
+from koala.cogs.twitch_alert.twitch_handler import TwitchAPIHandler
+from koala.cogs.twitch_alert.utils import TWITCH_KEY, TWITCH_SECRET
 import pytest
 
 @pytest.fixture
